Cherry pick commits from UHD 4.2 to 4.5
Previously, I branched from a stable branch 4.2  of ettus/uhd. We
recently upgraded to 4.5, but I cant merge due to loads of conflicts.
Instead, this is branched from master, so when a new release of ettus
comes out, the merge from upstream shouldnt cause any problems since it
should be a fast-forward